refactor(xtask): split main.rs into commands submodules and tests (#182)
Continues the context-burn reduction refactor series (see #184).

main.rs (1,870 lines) split into:
- main.rs         — CLI structs, dispatch, mod declarations (~110 lines)
- commands/mod.rs — re-exports all command functions
- commands/generate.rs  — generate_tests(), generate_code()
- commands/local_dev.rs — local_dev_setup/health/clean()
- commands/things.rs    — things_validate/backup/db_location()
- commands/analyze.rs   — analyze(), perf_test()
- commands/git_hooks.rs — setup_git_hooks()
- tests.rs        — all 48 tests (previously inline in main.rs)

No behavior changes.
