+++ tools/xtask/src/commands/mod.rs
@@ -1,0 +1,11 @@
+pub mod analyze;
+pub mod generate;
+pub mod git_hooks;
+pub mod local_dev;
+pub mod things;
+
+pub use analyze::{analyze, perf_test};
+pub use generate::{generate_code, generate_tests};
+pub use git_hooks::setup_git_hooks;
+pub use local_dev::{local_dev_clean, local_dev_health, local_dev_setup};
+pub use things::{things_backup, things_db_location, things_validate};
