+++ tools/xtask/src/main.rs
@@ -1,10 +1,13 @@
 //! Xtask - Build and development tools for Things 3 integration
 
+mod commands;
+use commands::{
+    analyze, generate_code, generate_tests, local_dev_clean, local_dev_health, local_dev_setup,
+    perf_test, setup_git_hooks, things_backup, things_db_location, things_validate,
+};
+
 use anyhow::Result;
 use clap::{Parser, Subcommand};
-use std::fs;
-use std::os::unix::fs::PermissionsExt;
-use std::path::Path;
 
 #[derive(Parser)]
 #[command(name = "xtask")]
@@ -111,1760 +114,5 @@ fn main() -> Result<()> {
     Ok(())
 }
 
-fn generate_tests(target: &str) {
-    println!("🔧 Generating test suite for: {target}");
-    println!("📝 This will create comprehensive unit tests");
-    println!("✅ Test generation complete!");
-}
-
-fn generate_code(code: &str) {
-    println!("🔧 Generating code: {code}");
-    println!("📝 This will create the requested code");
-    println!("✅ Code generation complete!");
-}
-
-fn local_dev_setup() {
-    println!("🚀 Setting up local development environment...");
-    println!("📦 Installing dependencies...");
-    println!("🔧 Configuring tools...");
-    println!("✅ Local development setup complete!");
-}
-
-fn local_dev_health() {
-    println!("🔍 Running health check...");
-    println!("✅ All systems healthy!");
-}
-
-fn local_dev_clean() {
-    println!("🧹 Cleaning build artifacts...");
-    println!("✅ Cleanup complete!");
-}
-
-fn things_validate() {
-    println!("🔍 Validating Things database...");
-    println!("✅ Database validation complete!");
-}
-
-fn things_backup() {
-    println!("💾 Backing up Things database...");
-    println!("✅ Backup complete!");
-}
-
-fn things_db_location() {
-    let db_path = things3_core::get_default_database_path();
-    println!("📁 Things database location: {}", db_path.display());
-}
-
-fn analyze() {
-    println!("🔍 Running code analysis...");
-    println!("✅ Analysis complete!");
-}
-
-fn perf_test() {
-    println!("⚡ Running performance tests...");
-    println!("✅ Performance tests complete!");
-}
-
-fn setup_git_hooks() -> Result<()> {
-    println!("🔧 Setting up git hooks...");
-
-    // Create .git/hooks directory if it doesn't exist
-    let hooks_dir = Path::new(".git/hooks");
-    if !hooks_dir.exists() {
-        fs::create_dir_all(hooks_dir)?;
-    }
-
-    // Create pre-commit hook
-    let pre_commit_hook = r#"#!/bin/bash
-# Pre-commit hook for Rust Things project
-
-echo "🔍 Running pre-commit checks..."
-
-# Format code
-echo "📝 Formatting code..."
-cargo fmt --all
-if [ $? -ne 0 ]; then
-    echo "❌ Code formatting failed"
-    exit 1
-fi
-
-# Run clippy with pedantic lints
-echo "🔍 Running clippy..."
-cargo clippy --all-targets --all-features -- -D warnings -D clippy::pedantic -A clippy::missing_docs_in_private_items -A clippy::module_name_repetitions
-if [ $? -ne 0 ]; then
-    echo "❌ Clippy checks failed"
-    exit 1
-fi
-
-# Run tests
-echo "🧪 Running tests..."
-cargo test --all-features
-if [ $? -ne 0 ]; then
-    echo "❌ Tests failed"
-    exit 1
-fi
-
-echo "✅ All pre-commit checks passed!"
-"#;
-
-    let pre_commit_path = hooks_dir.join("pre-commit");
-    fs::write(&pre_commit_path, pre_commit_hook)?;
-
-    // Make the hook executable
-    let mut perms = fs::metadata(&pre_commit_path)?.permissions();
-    perms.set_mode(0o755);
-    fs::set_permissions(&pre_commit_path, perms)?;
-
-    // Create pre-push hook
-    let pre_push_hook = r#"#!/bin/bash
-# Pre-push hook for Rust Things project
-
-echo "🔍 Running pre-push checks..."
-
-# Run clippy with pedantic lints
-echo "🔍 Running clippy..."
-cargo clippy --all-targets --all-features -- -D warnings -D clippy::pedantic -A clippy::missing_docs_in_private_items -A clippy::module_name_repetitions
-if [ $? -ne 0 ]; then
-    echo "❌ Clippy checks failed"
-    exit 1
-fi
-
-# Run tests
-echo "🧪 Running tests..."
-cargo test --all-features
-if [ $? -ne 0 ]; then
-    echo "❌ Tests failed"
-    exit 1
-fi
-
-echo "✅ All pre-push checks passed!"
-"#;
-
-    let pre_push_path = hooks_dir.join("pre-push");
-    fs::write(&pre_push_path, pre_push_hook)?;
-
-    // Make the hook executable
-    let mut perms = fs::metadata(&pre_push_path)?.permissions();
-    perms.set_mode(0o755);
-    fs::set_permissions(&pre_push_path, perms)?;
-
-    println!("✅ Git hooks installed successfully!");
-    println!("📝 Pre-commit hook: .git/hooks/pre-commit");
-    println!("📝 Pre-push hook: .git/hooks/pre-push");
-    println!();
-    println!("The hooks will run:");
-    println!("  • cargo fmt --all");
-    println!("  • cargo clippy --all-targets --all-features -- -D warnings -D clippy::pedantic");
-    println!("  • cargo test --all-features");
-
-    Ok(())
-}
-
-// Serialize all tests that mutate the process-global cwd.
 #[cfg(test)]
-static CWD_MUTEX: std::sync::Mutex<()> = std::sync::Mutex::new(());
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use clap::Parser;
-
-    #[test]
-    fn test_cli_parsing() {
-        // Test that CLI can be parsed without panicking
-        let cli = Cli::try_parse_from(["xtask", "analyze"]).unwrap();
-        assert!(matches!(cli.command, Commands::Analyze));
-
-        let cli = Cli::try_parse_from(["xtask", "perf-test"]).unwrap();
-        assert!(matches!(cli.command, Commands::PerfTest));
-
-        let cli = Cli::try_parse_from(["xtask", "setup-hooks"]).unwrap();
-        assert!(matches!(cli.command, Commands::SetupHooks));
-    }
-
-    #[test]
-    fn test_generate_tests_command() {
-        let cli = Cli::try_parse_from(["xtask", "generate-tests", "things3-core"]).unwrap();
-        if let Commands::GenerateTests { target } = cli.command {
-            assert_eq!(target, "things3-core");
-        } else {
-            panic!("Expected GenerateTests command");
-        }
-    }
-
-    #[test]
-    fn test_generate_code_command() {
-        let cli = Cli::try_parse_from(["xtask", "generate-code", "test"]).unwrap();
-        if let Commands::GenerateCode { code } = cli.command {
-            assert_eq!(code, "test");
-        } else {
-            panic!("Expected GenerateCode command");
-        }
-    }
-
-    #[test]
-    fn test_local_dev_commands() {
-        let cli = Cli::try_parse_from(["xtask", "local-dev", "setup"]).unwrap();
-        if let Commands::LocalDev { action } = cli.command {
-            assert!(matches!(action, LocalDevAction::Setup));
-        } else {
-            panic!("Expected LocalDev command");
-        }
-
-        let cli = Cli::try_parse_from(["xtask", "local-dev", "health"]).unwrap();
-        if let Commands::LocalDev { action } = cli.command {
-            assert!(matches!(action, LocalDevAction::Health));
-        } else {
-            panic!("Expected LocalDev command");
-        }
-
-        let cli = Cli::try_parse_from(["xtask", "local-dev", "clean"]).unwrap();
-        if let Commands::LocalDev { action } = cli.command {
-            assert!(matches!(action, LocalDevAction::Clean));
-        } else {
-            panic!("Expected LocalDev command");
-        }
-    }
-
-    #[test]
-    fn test_things_commands() {
-        let cli = Cli::try_parse_from(["xtask", "things", "validate"]).unwrap();
-        if let Commands::Things { action } = cli.command {
-            assert!(matches!(action, ThingsAction::Validate));
-        } else {
-            panic!("Expected Things command");
-        }
-
-        let cli = Cli::try_parse_from(["xtask", "things", "backup"]).unwrap();
-        if let Commands::Things { action } = cli.command {
-            assert!(matches!(action, ThingsAction::Backup));
-        } else {
-            panic!("Expected Things command");
-        }
-
-        let cli = Cli::try_parse_from(["xtask", "things", "db-location"]).unwrap();
-        if let Commands::Things { action } = cli.command {
-            assert!(matches!(action, ThingsAction::DbLocation));
-        } else {
-            panic!("Expected Things command");
-        }
-    }
-
-    #[test]
-    fn test_generate_tests_function() {
-        // Test that the function doesn't panic
-        generate_tests("test-target");
-    }
-
-    #[test]
-    fn test_generate_code_function() {
-        // Test that the function doesn't panic
-        generate_code("test-code");
-    }
-
-    #[test]
-    fn test_local_dev_setup_function() {
-        // Test that the function doesn't panic
-        local_dev_setup();
-    }
-
-    #[test]
-    fn test_local_dev_health_function() {
-        // Test that the function doesn't panic
-        local_dev_health();
-    }
-
-    #[test]
-    fn test_local_dev_clean_function() {
-        // Test that the function doesn't panic
-        local_dev_clean();
-    }
-
-    #[test]
-    fn test_things_validate_function() {
-        // Test that the function doesn't panic
-        things_validate();
-    }
-
-    #[test]
-    fn test_things_backup_function() {
-        // Test that the function doesn't panic
-        things_backup();
-    }
-
-    #[test]
-    fn test_things_db_location_function() {
-        // Test that the function doesn't panic
-        things_db_location();
-    }
-
-    #[test]
-    fn test_analyze_function() {
-        // Test that the function doesn't panic
-        analyze();
-    }
-
-    #[test]
-    fn test_perf_test_function() {
-        // Test that the function doesn't panic
-        perf_test();
-    }
-
-    #[test]
-    fn test_setup_git_hooks_function() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test that the function works with a temporary directory
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Create .git directory
-        if let Err(e) = std::fs::create_dir_all(".git/hooks") {
-            println!("Warning: Failed to create .git/hooks directory: {e:?}");
-            return;
-        }
-
-        // Test the function
-        let result = setup_git_hooks();
-        if result.is_err() {
-            // If it fails due to permission issues, that's okay for testing
-            // The important thing is that the function doesn't panic
-            println!("setup_git_hooks failed (expected in test environment): {result:?}");
-        } else {
-            // Verify hooks were created (only if they exist)
-            // In CI environments, the function might succeed but hooks might not be created
-            // due to permission issues or other constraints
-            let pre_commit_exists = std::path::Path::new(".git/hooks/pre-commit").exists();
-            let pre_push_exists = std::path::Path::new(".git/hooks/pre-push").exists();
-
-            if pre_commit_exists && !pre_push_exists {
-                // If pre-commit exists but pre-push doesn't, this might be a CI environment issue
-                println!("Warning: pre-commit hook exists but pre-push hook doesn't - this might be expected in CI");
-            } else if pre_commit_exists {
-                // Only assert if both should exist
-                assert!(
-                    pre_push_exists,
-                    "pre-push hook should exist if pre-commit hook exists"
-                );
-            }
-        }
-
-        // Restore original directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_setup_git_hooks_creates_directory() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test that the function creates the hooks directory if it doesn't exist
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Create .git directory first
-        if let Err(e) = std::fs::create_dir_all(".git") {
-            println!("Warning: Failed to create .git directory: {e:?}");
-            return;
-        }
-
-        // Test the function
-        let result = setup_git_hooks();
-
-        // Check the result and verify directory creation BEFORE changing back
-        match result {
-            Ok(()) => {
-                // Function succeeded, verify directory was created in temp directory
-                let current_dir =
-                    std::env::current_dir().unwrap_or_else(|_| std::path::PathBuf::from("."));
-                let git_path = std::path::Path::new(".git");
-                let hooks_path = std::path::Path::new(".git/hooks");
-                let abs_hooks_path = current_dir.join(".git/hooks");
-
-                // Check if either relative or absolute path exists
-                let hooks_exists = hooks_path.exists() || abs_hooks_path.exists();
-
-                if !hooks_exists {
-                    // Debug information
-                    println!("Current working directory: {current_dir:?}");
-                    println!(".git exists: {}", git_path.exists());
-                    println!("Checking if .git/hooks exists: {}", hooks_path.exists());
-                    println!(
-                        "Checking if absolute .git/hooks exists: {}",
-                        abs_hooks_path.exists()
-                    );
-
-                    if git_path.exists() {
-                        if let Ok(entries) = std::fs::read_dir(".git") {
-                            println!("Contents of .git directory:");
-                            for entry in entries.flatten() {
-                                println!("  {:?}", entry.path());
-                            }
-                        }
-                    }
-                }
-
-                // Only assert if the function succeeded and we're in a test environment where it should work
-                // In CI environments, the function might succeed but still not create the directory due to permissions
-                if hooks_exists {
-                    println!("✅ .git/hooks directory created successfully");
-                } else {
-                    println!("⚠️  .git/hooks directory not created, but this might be expected in CI environment");
-                    // Don't fail the test in CI environments where permissions might prevent directory creation
-                }
-            }
-            Err(e) => {
-                // Function failed, which might be expected in CI environment
-                println!("setup_git_hooks failed (expected in test environment): {e:?}");
-                // In CI environments, this might fail due to permissions or other issues
-                // We'll just log the error and continue
-            }
-        }
-
-        // Always restore original directory last
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_main_function_execution_paths() {
-        // Test that main function can be called with different commands
-        // This tests the main function execution paths that aren't covered by individual tests
-
-        // Test with analyze command
-        let args = ["xtask", "analyze"];
-        let cli = Cli::try_parse_from(args).unwrap();
-        match cli.command {
-            Commands::Analyze => {
-                // This path is covered
-            }
-            _ => panic!("Expected Analyze command"),
-        }
-
-        // Test with perf-test command
-        let args = ["xtask", "perf-test"];
-        let cli = Cli::try_parse_from(args).unwrap();
-        match cli.command {
-            Commands::PerfTest => {
-                // This path is covered
-            }
-            _ => panic!("Expected PerfTest command"),
-        }
-
-        // Test with setup-hooks command
-        let args = ["xtask", "setup-hooks"];
-        let cli = Cli::try_parse_from(args).unwrap();
-        match cli.command {
-            Commands::SetupHooks => {
-                // This path is covered
-            }
-            _ => panic!("Expected SetupHooks command"),
-        }
-    }
-
-    #[test]
-    fn test_things_db_location_with_env() {
-        // Test things_db_location function with different HOME environment
-        let original_home = std::env::var("HOME").ok();
-
-        // Test with custom HOME
-        std::env::set_var("HOME", "/custom/home");
-        things_db_location();
-
-        // Test with missing HOME (should use fallback)
-        std::env::remove_var("HOME");
-        things_db_location();
-
-        // Restore original HOME
-        if let Some(home) = original_home {
-            std::env::set_var("HOME", home);
-        } else {
-            std::env::remove_var("HOME");
-        }
-    }
-
-    #[test]
-    fn test_all_local_dev_actions() {
-        // Test all local dev action variants
-        let actions = [
-            ("setup", LocalDevAction::Setup),
-            ("health", LocalDevAction::Health),
-            ("clean", LocalDevAction::Clean),
-        ];
-
-        for (action_name, _expected_action) in actions {
-            let cli = Cli::try_parse_from(["xtask", "local-dev", action_name]).unwrap();
-            if let Commands::LocalDev { action } = cli.command {
-                assert!(matches!(action, _expected_action));
-            } else {
-                panic!("Expected LocalDev command for action: {action_name}");
-            }
-        }
-    }
-
-    #[test]
-    fn test_all_things_actions() {
-        // Test all things action variants
-        let actions = [
-            ("validate", ThingsAction::Validate),
-            ("backup", ThingsAction::Backup),
-            ("db-location", ThingsAction::DbLocation),
-        ];
-
-        for (action_name, _expected_action) in actions {
-            let cli = Cli::try_parse_from(["xtask", "things", action_name]).unwrap();
-            if let Commands::Things { action } = cli.command {
-                assert!(matches!(action, _expected_action));
-            } else {
-                panic!("Expected Things command for action: {action_name}");
-            }
-        }
-    }
-
-    #[test]
-    fn test_git_hooks_content() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test that the git hooks contain expected content
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Create .git directory
-        if let Err(e) = std::fs::create_dir_all(".git/hooks") {
-            println!("Warning: Failed to create .git/hooks directory: {e:?}");
-            return;
-        }
-
-        // Test the function
-        let result = setup_git_hooks();
-        if result.is_err() {
-            // If it fails due to permission issues, that's okay for testing
-            println!("setup_git_hooks failed (expected in test environment): {result:?}");
-            // Skip content verification if the function failed
-            println!("Skipping hook content verification due to function failure");
-        } else {
-            // Only verify content if the function succeeded
-            // Verify hooks exist and have executable permissions
-
-            // Check pre-commit hook
-            if let Ok(pre_commit_content) = std::fs::read_to_string(".git/hooks/pre-commit") {
-                // Check if hook has content (can be empty in some CI environments)
-                if pre_commit_content.is_empty() {
-                    println!(
-                        "⚠️  Warning: Pre-commit hook is empty (may be a CI environment issue)"
-                    );
-                    return; // Skip remaining checks
-                }
-
-                // Check if it's a rusty-hook wrapper or custom script
-                if pre_commit_content.contains("rusty-hook") {
-                    println!("✓ Pre-commit hook is managed by rusty-hook (.rusty-hook.toml)");
-                    assert!(pre_commit_content.contains("rusty-hook run"));
-                } else if pre_commit_content.contains("cargo") {
-                    println!("✓ Pre-commit hook is custom bash script");
-                    // Just verify it has cargo commands, don't be strict about exact format
-                    assert!(
-                        pre_commit_content.contains("cargo fmt")
-                            || pre_commit_content.contains("cargo clippy")
-                    );
-                } else {
-                    // In CI or other environments, hook format may vary
-                    println!(
-                        "Warning: Pre-commit hook has unexpected format (may be CI environment): {}",
-                        pre_commit_content
-                            .lines()
-                            .take(5)
-                            .collect::<Vec<_>>()
-                            .join("\n")
-                    );
-                    // Don't panic in test environment - just verify it's not empty
-                    assert!(!pre_commit_content.is_empty());
-                }
-            } else {
-                println!("Warning: Could not read pre-commit hook content");
-            }
-
-            // Check pre-push hook
-            if let Ok(pre_push_content) = std::fs::read_to_string(".git/hooks/pre-push") {
-                // In CI environments, the hook file may exist but be empty
-                if pre_push_content.is_empty() {
-                    println!("Warning: Pre-push hook exists but is empty (CI environment)");
-                } else if pre_push_content.contains("rusty-hook") {
-                    // Hook is managed by rusty-hook
-                    println!("✓ Pre-push hook is managed by rusty-hook (.rusty-hook.toml)");
-                    assert!(pre_push_content.contains("rusty-hook run"));
-                } else if pre_push_content.contains("cargo") {
-                    // Hook is a custom bash script
-                    println!("✓ Pre-push hook is custom bash script");
-                    // Just verify it has cargo commands, don't be strict about exact format
-                    assert!(
-                        pre_push_content.contains("cargo test")
-                            || pre_push_content.contains("cargo clippy")
-                    );
-                } else {
-                    // In CI or other environments, hook format may vary
-                    println!(
-                        "Warning: Pre-push hook has unexpected format (may be CI environment): {}",
-                        pre_push_content
-                            .lines()
-                            .take(5)
-                            .collect::<Vec<_>>()
-                            .join("\n")
-                    );
-                }
-            } else {
-                println!("Warning: Could not read pre-push hook content");
-            }
-        }
-
-        // Restore original directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_git_hooks_permissions() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test that git hooks are created with correct permissions
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Create .git directory
-        if let Err(e) = std::fs::create_dir_all(".git/hooks") {
-            println!("Warning: Failed to create .git/hooks directory: {e:?}");
-            return;
-        }
-
-        // Test the function
-        let result = setup_git_hooks();
-        if result.is_err() {
-            // If it fails due to permission issues, that's okay for testing
-            println!("setup_git_hooks failed (expected in test environment): {result:?}");
-        } else {
-            // Check permissions - only if files exist
-            if std::path::Path::new(".git/hooks/pre-commit").exists() {
-                if let Ok(pre_commit_metadata) = std::fs::metadata(".git/hooks/pre-commit") {
-                    // On Unix systems, check that the files are executable
-                    #[cfg(unix)]
-                    {
-                        use std::os::unix::fs::PermissionsExt;
-                        let pre_commit_perms = pre_commit_metadata.permissions();
-                        if pre_commit_perms.mode() & 0o111 == 0 {
-                            println!("Warning: Pre-commit hook is not executable");
-                        }
-                    }
-                } else {
-                    println!("Warning: Could not read pre-commit hook metadata");
-                }
-            } else {
-                println!("Warning: Pre-commit hook file does not exist");
-            }
-
-            if std::path::Path::new(".git/hooks/pre-push").exists() {
-                if let Ok(pre_push_metadata) = std::fs::metadata(".git/hooks/pre-push") {
-                    // On Unix systems, check that the files are executable
-                    #[cfg(unix)]
-                    {
-                        use std::os::unix::fs::PermissionsExt;
-                        let pre_push_perms = pre_push_metadata.permissions();
-                        if pre_push_perms.mode() & 0o111 == 0 {
-                            println!("Warning: Pre-push hook is not executable");
-                        }
-                    }
-                } else {
-                    println!("Warning: Could not read pre-push hook metadata");
-                }
-            } else {
-                println!("Warning: Pre-push hook file does not exist");
-            }
-        }
-
-        // Restore original directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_setup_git_hooks_creates_directory_when_missing() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test that the function creates the hooks directory when it doesn't exist
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Only create .git directory, not .git/hooks
-        if let Err(e) = std::fs::create_dir_all(".git") {
-            println!("Warning: Failed to create .git directory: {e:?}");
-            return;
-        }
-
-        // Capture absolute path before calling setup_git_hooks — avoids false negatives
-        // if a parallel test changes cwd between the call and the assertion.
-        let hooks_dir_abs = temp_dir.path().join(".git").join("hooks");
-
-        let result = setup_git_hooks();
-        if result.is_err() {
-            println!("setup_git_hooks failed (expected in some test environments): {result:?}");
-        } else {
-            assert!(
-                hooks_dir_abs.exists(),
-                "Expected .git/hooks directory to be created at {}, but it doesn't exist",
-                hooks_dir_abs.display()
-            );
-        }
-
-        // Restore original directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_things_db_location_with_no_home() {
-        // Test things_db_location function when HOME is not set
-        let original_home = std::env::var("HOME").ok();
-
-        // Remove HOME environment variable
-        std::env::remove_var("HOME");
-        things_db_location();
-
-        // Restore original HOME
-        if let Some(home) = original_home {
-            std::env::set_var("HOME", home);
-        } else {
-            std::env::remove_var("HOME");
-        }
-    }
-
-    #[test]
-    fn test_git_hooks_content_verification() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test that the git hooks content verification works when files exist
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Create .git directory
-        if let Err(e) = std::fs::create_dir_all(".git/hooks") {
-            println!("Warning: Failed to create .git/hooks directory: {e:?}");
-            return;
-        }
-
-        // Test the function
-        let result = setup_git_hooks();
-        if result.is_ok() {
-            // Test content verification paths - only if files exist
-            if std::path::Path::new(".git/hooks/pre-commit").exists() {
-                if let Ok(pre_commit_content) = std::fs::read_to_string(".git/hooks/pre-commit") {
-                    // Check for key content in the pre-commit hook - use soft checks
-                    if !pre_commit_content.contains("cargo fmt") {
-                        println!("Warning: Pre-commit hook missing cargo fmt");
-                    }
-                    if !pre_commit_content.contains("cargo clippy") {
-                        println!("Warning: Pre-commit hook missing cargo clippy");
-                    }
-                    if !pre_commit_content.contains("cargo test") {
-                        println!("Warning: Pre-commit hook missing cargo test");
-                    }
-                } else {
-                    println!("Warning: Could not read pre-commit hook content");
-                }
-            } else {
-                println!("Warning: Pre-commit hook file does not exist");
-            }
-
-            if std::path::Path::new(".git/hooks/pre-push").exists() {
-                if let Ok(pre_push_content) = std::fs::read_to_string(".git/hooks/pre-push") {
-                    // Check for key content in the pre-push hook - use soft checks
-                    if !pre_push_content.contains("cargo clippy") {
-                        println!("Warning: Pre-push hook missing cargo clippy");
-                    }
-                    if !pre_push_content.contains("cargo test") {
-                        println!("Warning: Pre-push hook missing cargo test");
-                    }
-                } else {
-                    println!("Warning: Could not read pre-push hook content");
-                }
-            } else {
-                println!("Warning: Pre-push hook file does not exist");
-            }
-        } else {
-            println!("Warning: setup_git_hooks failed: {result:?}");
-        }
-
-        // Restore original directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_git_hooks_permissions_error_path() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test the error handling path in git hooks permissions test
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Create .git directory
-        if let Err(e) = std::fs::create_dir_all(".git/hooks") {
-            println!("Warning: Failed to create .git/hooks directory: {e:?}");
-            return;
-        }
-
-        // Test the function
-        let result = setup_git_hooks();
-        if result.is_err() {
-            // This should trigger the error handling path in the test
-            println!("setup_git_hooks failed (expected in test environment): {result:?}");
-        }
-
-        // Restore original directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_setup_git_hooks_error_handling() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test error handling paths in setup_git_hooks function
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = match std::env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                println!("Warning: Failed to get current directory: {e:?}");
-                return;
-            }
-        };
-
-        // Change to temp directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(temp_dir.path()) {
-            println!("Warning: Failed to change to temp directory: {e:?}");
-            return;
-        }
-
-        // Create .git directory but make it read-only to force errors
-        if let Err(e) = std::fs::create_dir_all(".git/hooks") {
-            println!("Warning: Failed to create .git/hooks directory: {e:?}");
-            return;
-        }
-
-        // Make the hooks directory read-only (this might not work on all systems)
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            if let Ok(metadata) = std::fs::metadata(".git/hooks") {
-                let mut perms = metadata.permissions();
-                perms.set_mode(0o444); // Read-only
-                let _ = std::fs::set_permissions(".git/hooks", perms);
-            }
-        }
-
-        // Test the function - this should trigger error paths
-        let result = setup_git_hooks();
-        if result.is_err() {
-            // This should trigger the error handling paths in the function
-            println!("setup_git_hooks failed as expected: {result:?}");
-        }
-
-        // Restore original directory - handle potential errors gracefully
-        if let Err(e) = std::env::set_current_dir(&original_dir) {
-            println!("Warning: Failed to restore original directory: {e:?}");
-        }
-    }
-
-    #[test]
-    fn test_main_function_with_setup_hooks() {
-        // Test the main function execution path for setup-hooks command
-        // This tests the main function match statement
-        let cli = Cli::try_parse_from(["xtask", "setup-hooks"]).unwrap();
-        match cli.command {
-            Commands::SetupHooks => {
-                // This path is covered
-                println!("SetupHooks command parsed successfully");
-            }
-            _ => panic!("Expected SetupHooks command"),
-        }
-    }
-
-    #[test]
-    fn test_analyze_command_execution() {
-        let cli = Cli::try_parse_from(["xtask", "analyze"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::Analyze => {
-                analyze();
-            }
-            _ => panic!("Expected Analyze command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with analyze command"
-        );
-    }
-
-    #[test]
-    fn test_perf_test_command_execution() {
-        let cli = Cli::try_parse_from(["xtask", "perf-test"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::PerfTest => {
-                perf_test();
-            }
-            _ => panic!("Expected PerfTest command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with perf-test command"
-        );
-    }
-
-    #[test]
-    fn test_generate_tests_command_execution() {
-        let cli = Cli::try_parse_from(["xtask", "generate-tests", "test-target"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::GenerateTests { target } => {
-                generate_tests(&target);
-            }
-            _ => panic!("Expected GenerateTests command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with generate-tests command"
-        );
-    }
-
-    #[test]
-    fn test_generate_code_command_execution() {
-        let cli = Cli::try_parse_from(["xtask", "generate-code", "test-code"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::GenerateCode { code } => {
-                generate_code(&code);
-            }
-            _ => panic!("Expected GenerateCode command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with generate-code command"
-        );
-    }
-
-    #[test]
-    fn test_local_dev_commands_execution() {
-        // Test with local-dev setup command
-        let cli = Cli::try_parse_from(["xtask", "local-dev", "setup"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::LocalDev { action } => match action {
-                LocalDevAction::Setup => {
-                    local_dev_setup();
-                }
-                _ => panic!("Expected Setup action"),
-            },
-            _ => panic!("Expected LocalDev command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with local-dev setup command"
-        );
-
-        // Test with local-dev health command
-        let cli = Cli::try_parse_from(["xtask", "local-dev", "health"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::LocalDev { action } => match action {
-                LocalDevAction::Health => {
-                    local_dev_health();
-                }
-                _ => panic!("Expected Health action"),
-            },
-            _ => panic!("Expected LocalDev command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with local-dev health command"
-        );
-
-        // Test with local-dev clean command
-        let cli = Cli::try_parse_from(["xtask", "local-dev", "clean"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::LocalDev { action } => match action {
-                LocalDevAction::Clean => {
-                    local_dev_clean();
-                }
-                _ => panic!("Expected Clean action"),
-            },
-            _ => panic!("Expected LocalDev command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with local-dev clean command"
-        );
-    }
-
-    #[test]
-    fn test_things_commands_execution() {
-        // Test with things validate command
-        let cli = Cli::try_parse_from(["xtask", "things", "validate"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::Things { action } => match action {
-                ThingsAction::Validate => {
-                    things_validate();
-                }
-                _ => panic!("Expected Validate action"),
-            },
-            _ => panic!("Expected Things command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with things validate command"
-        );
-
-        // Test with things backup command
-        let cli = Cli::try_parse_from(["xtask", "things", "backup"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::Things { action } => match action {
-                ThingsAction::Backup => {
-                    things_backup();
-                }
-                _ => panic!("Expected Backup action"),
-            },
-            _ => panic!("Expected Things command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with things backup command"
-        );
-
-        // Test with things db-location command
-        let cli = Cli::try_parse_from(["xtask", "things", "db-location"]).unwrap();
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::Things { action } => match action {
-                ThingsAction::DbLocation => {
-                    things_db_location();
-                }
-                _ => panic!("Expected DbLocation action"),
-            },
-            _ => panic!("Expected Things command"),
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should not panic with things db-location command"
-        );
-    }
-
-    #[test]
-    fn test_main_function_error_handling() {
-        // Test that main function handles errors gracefully
-        // This tests the error handling paths in the main function
-
-        // Test with setup-hooks command that might fail
-        let cli = Cli::try_parse_from(["xtask", "setup-hooks"]).unwrap();
-        let result = std::panic::catch_unwind(|| {
-            match cli.command {
-                Commands::SetupHooks => {
-                    // This might fail in test environment, but should not panic
-                    let _ = setup_git_hooks();
-                }
-                _ => panic!("Expected SetupHooks command"),
-            }
-        });
-        assert!(
-            result.is_ok(),
-            "Main function should handle setup-hooks errors gracefully"
-        );
-    }
-
-    #[test]
-    fn test_main_function_comprehensive() {
-        // Test comprehensive main function execution with all command types
-        // This provides maximum coverage of the main function
-
-        let commands = [
-            ("analyze", Commands::Analyze),
-            ("perf-test", Commands::PerfTest),
-            (
-                "generate-tests",
-                Commands::GenerateTests {
-                    target: "test".to_string(),
-                },
-            ),
-            (
-                "generate-code",
-                Commands::GenerateCode {
-                    code: "test".to_string(),
-                },
-            ),
-            (
-                "local-dev",
-                Commands::LocalDev {
-                    action: LocalDevAction::Setup,
-                },
-            ),
-            (
-                "things",
-                Commands::Things {
-                    action: ThingsAction::Validate,
-                },
-            ),
-            ("setup-hooks", Commands::SetupHooks),
-        ];
-
-        for (cmd_name, _expected_command) in commands {
-            let args = match cmd_name {
-                "generate-tests" => vec!["xtask", "generate-tests", "test"],
-                "generate-code" => vec!["xtask", "generate-code", "test"],
-                "local-dev" => vec!["xtask", "local-dev", "setup"],
-                "things" => vec!["xtask", "things", "validate"],
-                _ => vec!["xtask", cmd_name],
-            };
-
-            let cli = Cli::try_parse_from(args).unwrap();
-            let result = std::panic::catch_unwind(|| match cli.command {
-                Commands::Analyze => analyze(),
-                Commands::PerfTest => perf_test(),
-                Commands::GenerateTests { target } => generate_tests(&target),
-                Commands::GenerateCode { code } => generate_code(&code),
-                Commands::LocalDev { action } => match action {
-                    LocalDevAction::Setup => local_dev_setup(),
-                    LocalDevAction::Health => local_dev_health(),
-                    LocalDevAction::Clean => local_dev_clean(),
-                },
-                Commands::Things { action } => match action {
-                    ThingsAction::Validate => things_validate(),
-                    ThingsAction::Backup => things_backup(),
-                    ThingsAction::DbLocation => things_db_location(),
-                },
-                Commands::SetupHooks => {
-                    let _ = setup_git_hooks();
-                }
-            });
-
-            assert!(
-                result.is_ok(),
-                "Main function should not panic with {cmd_name} command"
-            );
-        }
-    }
-
-    #[test]
-    fn test_cli_error_handling() {
-        // Test CLI parsing with invalid arguments
-        let result = Cli::try_parse_from(["xtask", "invalid-command"]);
-        assert!(result.is_err(), "Should fail with invalid command");
-
-        let result = Cli::try_parse_from(["xtask", "generate-tests"]);
-        assert!(
-            result.is_err(),
-            "Should fail when missing required argument"
-        );
-
-        let result = Cli::try_parse_from(["xtask", "generate-code"]);
-        assert!(
-            result.is_err(),
-            "Should fail when missing required argument"
-        );
-
-        let result = Cli::try_parse_from(["xtask", "local-dev"]);
-        assert!(result.is_err(), "Should fail when missing subcommand");
-
-        let result = Cli::try_parse_from(["xtask", "things"]);
-        assert!(result.is_err(), "Should fail when missing subcommand");
-    }
-
-    #[test]
-    fn test_main_function_with_actual_execution() {
-        // Test the main function by actually calling it with different commands
-        // This ensures the main function's match arms are all covered
-
-        // Test with analyze command
-        let result = std::panic::catch_unwind(|| {
-            let cli = Cli {
-                command: Commands::Analyze,
-            };
-            match cli.command {
-                Commands::Analyze => analyze(),
-                _ => unreachable!(),
-            }
-        });
-        assert!(result.is_ok());
-
-        // Test with perf-test command
-        let result = std::panic::catch_unwind(|| {
-            let cli = Cli {
-                command: Commands::PerfTest,
-            };
-            match cli.command {
-                Commands::PerfTest => perf_test(),
-                _ => unreachable!(),
-            }
-        });
-        assert!(result.is_ok());
-
-        // Test with generate-tests command
-        let result = std::panic::catch_unwind(|| {
-            let cli = Cli {
-                command: Commands::GenerateTests {
-                    target: "test-target".to_string(),
-                },
-            };
-            match cli.command {
-                Commands::GenerateTests { target } => generate_tests(&target),
-                _ => unreachable!(),
-            }
-        });
-        assert!(result.is_ok());
-
-        // Test with generate-code command
-        let result = std::panic::catch_unwind(|| {
-            let cli = Cli {
-                command: Commands::GenerateCode {
-                    code: "test-code".to_string(),
-                },
-            };
-            match cli.command {
-                Commands::GenerateCode { code } => generate_code(&code),
-                _ => unreachable!(),
-            }
-        });
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_setup_git_hooks_file_write_errors() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test error handling when file writing fails
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = std::env::current_dir().unwrap_or_default();
-
-        // Change to temp directory
-        if std::env::set_current_dir(temp_dir.path()).is_err() {
-            return; // Skip test if we can't change directory
-        }
-
-        // Create .git directory but make it read-only to force write errors
-        if std::fs::create_dir_all(".git/hooks").is_ok() {
-            // Make the hooks directory read-only on Unix systems
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                if let Ok(metadata) = std::fs::metadata(".git/hooks") {
-                    let mut perms = metadata.permissions();
-                    perms.set_mode(0o444); // Read-only
-                    let _ = std::fs::set_permissions(".git/hooks", perms);
-                }
-            }
-
-            // Test the function - this should trigger error paths
-            let result = setup_git_hooks();
-            // The function might succeed or fail depending on the system
-            // The important thing is that it doesn't panic
-            match result {
-                Ok(()) => println!("setup_git_hooks succeeded despite read-only directory"),
-                Err(e) => println!("setup_git_hooks failed as expected: {e:?}"),
-            }
-        }
-
-        // Restore original directory
-        let _ = std::env::set_current_dir(&original_dir);
-    }
-
-    #[test]
-    fn test_setup_git_hooks_permission_errors() {
-        let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        // Test permission error handling in setup_git_hooks
-        let temp_dir = tempfile::tempdir().unwrap();
-        let original_dir = std::env::current_dir().unwrap_or_default();
-
-        // Change to temp directory
-        if std::env::set_current_dir(temp_dir.path()).is_err() {
-            return; // Skip test if we can't change directory
-        }
-
-        // Create .git/hooks directory
-        if std::fs::create_dir_all(".git/hooks").is_ok() {
-            // Create a file with the same name as our hook to cause a conflict
-            let _ = std::fs::write(".git/hooks/pre-commit", "existing content");
-
-            // Make the file read-only to cause permission errors
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                if let Ok(metadata) = std::fs::metadata(".git/hooks/pre-commit") {
-                    let mut perms = metadata.permissions();
-                    perms.set_mode(0o444); // Read-only
-                    let _ = std::fs::set_permissions(".git/hooks/pre-commit", perms);
-                }
-            }
-
-            // Test the function
-            let result = setup_git_hooks();
-            // The function might succeed or fail depending on permissions
-            match result {
-                Ok(()) => println!("setup_git_hooks succeeded despite permission issues"),
-                Err(e) => println!("setup_git_hooks failed due to permissions: {e:?}"),
-            }
-        }
-
-        // Restore original directory
-        let _ = std::env::set_current_dir(&original_dir);
-    }
-
-    #[test]
-    fn test_edge_cases_and_error_paths() {
-        // Test various edge cases to improve coverage
-
-        // Test things_db_location with various HOME values
-        let original_home = std::env::var("HOME").ok();
-
-        // Test with empty HOME
-        std::env::set_var("HOME", "");
-        things_db_location();
-
-        // Test with HOME containing special characters
-        std::env::set_var("HOME", "/path/with spaces/and-dashes");
-        things_db_location();
-
-        // Test with very long HOME path
-        std::env::set_var(
-            "HOME",
-            "/very/long/path/that/goes/on/and/on/and/on/to/test/edge/cases",
-        );
-        things_db_location();
-
-        // Restore original HOME
-        match original_home {
-            Some(home) => std::env::set_var("HOME", home),
-            None => std::env::remove_var("HOME"),
-        }
-
-        // Test all function variants with different inputs
-        generate_tests("complex-target-name-with-dashes");
-        generate_code("complex_code_with_underscores");
-
-        // Test functions multiple times to ensure they're idempotent
-        for _ in 0..3 {
-            analyze();
-            perf_test();
-            local_dev_setup();
-            local_dev_health();
-            local_dev_clean();
-            things_validate();
-            things_backup();
-        }
-    }
-}
-
-#[test]
-fn test_main_function_all_commands() {
-    // Test that main function can handle all command types
-    // This provides comprehensive coverage of the main function
-
-    // Test generate-tests command
-    let args = ["xtask", "generate-tests", "test-target"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::GenerateTests { target } => {
-            assert_eq!(target, "test-target");
-        }
-        _ => panic!("Expected GenerateTests command"),
-    }
-
-    // Test generate-code command
-    let args = ["xtask", "generate-code", "test-code"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::GenerateCode { code } => {
-            assert_eq!(code, "test-code");
-        }
-        _ => panic!("Expected GenerateCode command"),
-    }
-
-    // Test local-dev setup command
-    let args = ["xtask", "local-dev", "setup"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::LocalDev {
-            action: LocalDevAction::Setup,
-        } => {
-            // This path is covered
-        }
-        _ => panic!("Expected LocalDev Setup command"),
-    }
-
-    // Test local-dev health command
-    let args = ["xtask", "local-dev", "health"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::LocalDev {
-            action: LocalDevAction::Health,
-        } => {
-            // This path is covered
-        }
-        _ => panic!("Expected LocalDev Health command"),
-    }
-
-    // Test local-dev clean command
-    let args = ["xtask", "local-dev", "clean"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::LocalDev {
-            action: LocalDevAction::Clean,
-        } => {
-            // This path is covered
-        }
-        _ => panic!("Expected LocalDev Clean command"),
-    }
-
-    // Test things validate command
-    let args = ["xtask", "things", "validate"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::Things {
-            action: ThingsAction::Validate,
-        } => {
-            // This path is covered
-        }
-        _ => panic!("Expected Things Validate command"),
-    }
-
-    // Test things backup command
-    let args = ["xtask", "things", "backup"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::Things {
-            action: ThingsAction::Backup,
-        } => {
-            // This path is covered
-        }
-        _ => panic!("Expected Things Backup command"),
-    }
-
-    // Test things db-location command
-    let args = ["xtask", "things", "db-location"];
-    let cli = Cli::try_parse_from(args).unwrap();
-    match cli.command {
-        Commands::Things {
-            action: ThingsAction::DbLocation,
-        } => {
-            // This path is covered
-        }
-        _ => panic!("Expected Things DbLocation command"),
-    }
-}
-
-#[test]
-fn test_main_function_result_handling() {
-    // Test that main function properly handles Result types
-
-    // Test successful execution path
-    let result = std::panic::catch_unwind(|| {
-        let cli = Cli {
-            command: Commands::Analyze,
-        };
-
-        // Simulate the main function logic
-        match cli.command {
-            Commands::Analyze => {
-                analyze();
-                Ok::<(), anyhow::Error>(())
-            }
-            _ => Ok(()),
-        }
-    });
-
-    assert!(result.is_ok());
-
-    // Test with setup-hooks command that returns Result
-    let result = std::panic::catch_unwind(|| {
-        let cli = Cli {
-            command: Commands::SetupHooks,
-        };
-
-        // Simulate the main function logic with Result handling
-        match cli.command {
-            Commands::SetupHooks => {
-                // This might fail, but should not panic
-                let _ = setup_git_hooks();
-                Ok::<(), anyhow::Error>(())
-            }
-            _ => Ok(()),
-        }
-    });
-
-    assert!(result.is_ok());
-}
-
-#[test]
-fn test_all_enum_variants_coverage() {
-    // Ensure all enum variants are covered in tests
-
-    // Test all Commands variants
-    let commands = [
-        Commands::Analyze,
-        Commands::PerfTest,
-        Commands::GenerateTests {
-            target: "test".to_string(),
-        },
-        Commands::GenerateCode {
-            code: "test".to_string(),
-        },
-        Commands::LocalDev {
-            action: LocalDevAction::Setup,
-        },
-        Commands::LocalDev {
-            action: LocalDevAction::Health,
-        },
-        Commands::LocalDev {
-            action: LocalDevAction::Clean,
-        },
-        Commands::Things {
-            action: ThingsAction::Validate,
-        },
-        Commands::Things {
-            action: ThingsAction::Backup,
-        },
-        Commands::Things {
-            action: ThingsAction::DbLocation,
-        },
-        Commands::SetupHooks,
-    ];
-
-    for command in commands {
-        let cli = Cli { command };
-
-        // Test that each command can be matched without panicking
-        let result = std::panic::catch_unwind(|| match cli.command {
-            Commands::Analyze => analyze(),
-            Commands::PerfTest => perf_test(),
-            Commands::GenerateTests { target } => generate_tests(&target),
-            Commands::GenerateCode { code } => generate_code(&code),
-            Commands::LocalDev { action } => match action {
-                LocalDevAction::Setup => local_dev_setup(),
-                LocalDevAction::Health => local_dev_health(),
-                LocalDevAction::Clean => local_dev_clean(),
-            },
-            Commands::Things { action } => match action {
-                ThingsAction::Validate => things_validate(),
-                ThingsAction::Backup => things_backup(),
-                ThingsAction::DbLocation => things_db_location(),
-            },
-            Commands::SetupHooks => {
-                let _ = setup_git_hooks();
-            }
-        });
-
-        assert!(result.is_ok(), "Command execution should not panic");
-    }
-}
-
-#[test]
-fn test_setup_git_hooks_directory_creation_edge_cases() {
-    let _cwd_guard = CWD_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-    // Test edge cases in directory creation
-    let temp_dir = tempfile::tempdir().unwrap();
-    let original_dir = std::env::current_dir().unwrap_or_default();
-
-    // Change to temp directory
-    if std::env::set_current_dir(temp_dir.path()).is_err() {
-        return;
-    }
-
-    // Test when .git exists but hooks doesn't
-    if std::fs::create_dir_all(".git").is_ok() {
-        // Ensure hooks directory doesn't exist
-        let _ = std::fs::remove_dir_all(".git/hooks");
-
-        // Test the function
-        let result = setup_git_hooks();
-        match result {
-            Ok(()) => {
-                // Verify directory was created
-                assert!(
-                    std::path::Path::new(".git/hooks").exists() || std::env::var("CI").is_ok(), // Allow CI environments to behave differently
-                    "Hooks directory should be created"
-                );
-            }
-            Err(e) => {
-                println!("setup_git_hooks failed: {e:?}");
-                // In some environments, this might fail due to permissions
-            }
-        }
-    }
-
-    // Restore original directory
-    let _ = std::env::set_current_dir(&original_dir);
-}
-
-#[test]
-fn test_function_output_and_behavior() {
-    // Test that functions produce expected output patterns
-    // These functions print to stdout, so we test they don't panic and complete successfully
-
-    let functions_to_test = [
-        || generate_tests("output-test"),
-        || generate_code("output-test"),
-        || local_dev_setup(),
-        || local_dev_health(),
-        || local_dev_clean(),
-        || things_validate(),
-        || things_backup(),
-        || things_db_location(),
-        || analyze(),
-        || perf_test(),
-    ];
-
-    for (i, func) in functions_to_test.iter().enumerate() {
-        let result = std::panic::catch_unwind(func);
-        assert!(result.is_ok(), "Function {} should not panic", i);
-    }
-}
-
-#[test]
-fn test_cli_struct_construction() {
-    // Test CLI struct can be constructed with all command variants
-    let test_cases = [
-        Cli {
-            command: Commands::Analyze,
-        },
-        Cli {
-            command: Commands::PerfTest,
-        },
-        Cli {
-            command: Commands::SetupHooks,
-        },
-        Cli {
-            command: Commands::GenerateTests {
-                target: "test-construction".to_string(),
-            },
-        },
-        Cli {
-            command: Commands::GenerateCode {
-                code: "test-construction".to_string(),
-            },
-        },
-        Cli {
-            command: Commands::LocalDev {
-                action: LocalDevAction::Setup,
-            },
-        },
-        Cli {
-            command: Commands::LocalDev {
-                action: LocalDevAction::Health,
-            },
-        },
-        Cli {
-            command: Commands::LocalDev {
-                action: LocalDevAction::Clean,
-            },
-        },
-        Cli {
-            command: Commands::Things {
-                action: ThingsAction::Validate,
-            },
-        },
-        Cli {
-            command: Commands::Things {
-                action: ThingsAction::Backup,
-            },
-        },
-        Cli {
-            command: Commands::Things {
-                action: ThingsAction::DbLocation,
-            },
-        },
-    ];
-
-    // Test that all CLI structs can be constructed and used
-    for cli in test_cases {
-        // Just verify the struct is valid by accessing its command field
-        match cli.command {
-            Commands::Analyze => {}
-            Commands::PerfTest => {}
-            Commands::SetupHooks => {}
-            Commands::GenerateTests { target: _ } => {}
-            Commands::GenerateCode { code: _ } => {}
-            Commands::LocalDev { action: _ } => {}
-            Commands::Things { action: _ } => {}
-        }
-    }
-}
+mod tests;
